requirements.txt
flask==3.0.0
requests==2.32.4

# Development/Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-playwright>=0.4.0
